+++ Editor/Models/VRChat/AvatarDynamics.cs
@@ -63,11 +63,28 @@ namespace KRT.VRCQuestTools.Models.VRChat
             {
                 var transformCount = CalculatePhysBoneTransformCount(pb) - 1; // ignore itself.
                 var rootTrans = pb.RootTransform == null ? pb.GameObject.transform : pb.RootTransform;
-                var childCount = rootTrans.childCount - pb.IgnoreTransforms.Where(t => t != null).Count(t => t.IsChildOf(rootTrans)); // count children without ignored transforms.
-                if (childCount > 1)
+
+                var multiChildRoots = rootTrans.GetComponentsInChildren<Transform>(true)
+                    .Where(t => !IsFinallyEditorOnly(root, t.gameObject))
+                    .Where(t => IsMultiChildRoot(pb, t))
+                    .ToArray();
+                transformCount -= multiChildRoots.Sum(t => t.childCount);
+
+                if (pb.MultiChildType != VRCSDKUtility.Reflection.PhysBone.MultiChildTypeEnum.Ignore)
                 {
-                    transformCount -= childCount; // ignore children's first objects.
+                    transformCount += multiChildRoots.Length;
                 }
+
+                if (pb.EndpointPosition.magnitude > 0)
+                {
+                    var ignore = pb.IgnoreTransforms.ToArray();
+                    var endpoints = rootTrans.GetComponentsInChildren<Transform>(true)
+                        .Where(t => t.childCount == 0)
+                        .Where(t => !IsIgnoredTransform(t, ignore))
+                        .Where(t => !IsFinallyEditorOnly(root, t.gameObject));
+                    transformCount += endpoints.Count();
+                }
+
                 var colliderCount = pb.Colliders
                     .Distinct()
                     .Where(c => c != null)
@@ -105,6 +122,29 @@ namespace KRT.VRCQuestTools.Models.VRChat
                 return false;
             }
             return IsFinallyEditorOnly(root, obj.transform.parent.gameObject);
+        }
+
+        private static bool IsIgnoredTransform(Transform transform, Transform[] ignoreTransforms)
+        {
+            ignoreTransforms
+                .Where(t => t != null)
+                .SelectMany(t => t.GetComponentsInChildren<Transform>(true))
+                .ToArray();
+            return ignoreTransforms.Contains(transform);
+        }
+
+        private static bool IsMultiChildRoot(VRCSDKUtility.Reflection.PhysBone physBone, Transform transform)
+        {
+            if (transform.childCount <= 1)
+            {
+                return false;
+            }
+
+            // get all direct child transform.
+            var children = transform.GetComponentsInChildren<Transform>(true)
+                .Where(t => t.parent == transform)
+                .ToArray();
+            return children.Where(t => !IsIgnoredTransform(t, physBone.IgnoreTransforms.ToArray())).Count() > 1;
         }
 
         private static int CalculatePhysBoneTransformCount(VRCSDKUtility.Reflection.PhysBone physbone)